refactor: consolidate mutation views
Amp-Thread-ID: https://ampcode.com/threads/T-01a01cf7-d051-7436-af1a-79973166e190
